long

+++ src/Domain/Repositories/IProductRepository.cs
@@ -4,8 +4,8 @@ namespace Domain.Repositories;
 
 public interface IProductRepository : IGenericRepository<Product>
 {
-    Task<bool> HasSufficientStockAsync(int productId, int qty);
+    Task<bool> HasSufficientStockAsync(long productId, long qty);
 
-    Task<bool> ReduceStockQtyAsync(int productId, int qty);
+    Task<bool> ReduceStockQtyAsync(long productId, long qty);
 
 }
