+++ src/Infrastructure/Repositories/ProductRepository.cs
@@ -8,13 +8,13 @@ namespace Infrastructure.Repositories;
 public class ProductRepository(AppDbContext context)
     : GenericRepository<Product>(context), IProductRepository
 {
-    public async Task<bool> HasSufficientStockAsync(int productId, int qty)
+    public async Task<bool> HasSufficientStockAsync(long productId, long qty)
     {
         var product = await Context.Products.FindAsync(productId);
         return product is not null && product.Stock >= qty;
     }
 
-    public async Task<bool> ReduceStockQtyAsync(int productId, int qty)
+    public async Task<bool> ReduceStockQtyAsync(long productId, long qty)
     {
         var product = await Context.Products.FindAsync(productId);
         if (product is null || product.Stock < qty) return false;
